+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- Only 'id', 'first_name', 'last_name' required in user info JSON
+- only 'id' required in place info JSON attached to user info
+
 ## [Version 1.2.1, Build 650] - 2019-12-04
 
 ### Changed