+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [Version 1.2.2, Build 651] - 2019-12-08
+
 ### Fixed
 - Only 'id', 'first_name', 'last_name' and location/country id' required in user info JSON
 - Hotel brand names now loaded from server